Add additional unit tests. Found bug with multiplication and division that needs fixing.
