+++ Calculator/Model/Input.swift
@@ -37,70 +37,6 @@ enum Input: Hashable {
     
 }
 
-// MARK: - Instruction
-// MARK: -
-extension Input {
-    enum Instruction: Hashable, Displayable {
-        case `operator`(Operator)
-        case stateChange(StateChange)
-        
-        var displayValue: String {
-            switch self {
-            case .operator(let `operator`): return `operator`.displayValue
-            case .stateChange(let stateChange): return stateChange.displayValue
-            }
-        }
-        
-        var `operator`: Operator? {
-            switch self {
-            case .`operator`(let `operator`): return `operator`
-            case .stateChange: return nil
-            }
-        }
-        
-        var stateChange: StateChange? {
-            switch self {
-            case .`operator`: return nil
-            case .stateChange(let stateChange): return stateChange
-            }
-        }
-    }
-}
-
-// MARK: - Operator
-// MARK: -
-extension Input {
-    enum Operator: Hashable, CustomStringConvertible, Displayable {
-        case binaryOperator(BinaryOperator)
-        case unaryOperator(UnaryOperator)
-        
-        /// Equal is a very special operator, because most calculators should
-        /// "cache" the last binary operator (if any) and the last operand (if any)
-        /// and perform the operator several times to the result, using the last
-        /// operand.
-        case equal
-        
-        enum BinaryOperator: String, CaseIterable, Hashable, CustomStringConvertible, Displayable {
-            case addition, subtaction, multiplication, division
-        }
-        enum UnaryOperator: String, CaseIterable, Hashable, CustomStringConvertible, Displayable {
-            case negation, percent
-        }
-    }
-    
-    enum StateChange: String, Hashable, CustomStringConvertible, Displayable {
-        
-        /// Inserts a comma, making a presented integer into a presented float.
-        case comma
-        
-        /// Clear or "All Clear", a.k.a. "AC"
-        case clear
-        
-        /// Saves the number, a.k.a. "Memory".
-        case saveNumber
-    }
-}
-
 extension Input.StateChange {
     var displayValue: String {
         switch self {
@@ -109,41 +45,6 @@ extension Input.StateChange {
         case .comma: return Locale.current.decimalSeparator ?? "."
         }
     }
-}
-
-extension Input.Operator.BinaryOperator {
-    var displayValue: String {
-        switch self {
-        case .addition: return "+"
-        case .subtaction: return "-"
-        case .multiplication: return "X"
-        case .division: return "/"
-        }
-    }
-}
-
-extension Input.Operator.UnaryOperator {
-    var displayValue: String {
-        switch self {
-        case .percent: return "%"
-        case .negation: return "+/-"
-        }
-    }
-}
-
-extension Input.Operator {
-    
-    var displayValue: String {
-        switch self {
-        case .equal: return "="
-        case .binaryOperator(let binaryOperator):
-            return binaryOperator.displayValue
-        case .unaryOperator(let unaryOperator):
-            return unaryOperator.displayValue
-        }
-    }
-    
-    
 }
 
 // MARK: - Digit
