+++ CalculatorTests/CalculatorTests.swift
@@ -4,42 +4,30 @@ import XCTest
 final class CalculatorTests: XCTestCase {
 
     func test_instruction_equals_add3() {
-        let calculator = Calculator._withState(
-            value: 3,
-            lastBinaryOperator: .addition
-//            ,
-//            newOperand: 3
-        )
+        let calculator = Calculator()
        
-        func assertInternalState() {
-//             XCTAssertEqual(calculator._lastOperand, 3)
-//             XCTAssertEqual(calculator._lastBinaryOperator, .addition)
-//             XCTAssertTrue(calculator._digits.isEmpty)
+        func initialState() {
+            calculator.input(2)
+            calculator.input("+")
+            calculator.input(3)
         }
         
-        assertInternalState()
+        initialState()
 
         XCTAssertEqual(
             calculator.input(.instruction(.operator(.equal))),
-            6
+            5
         )
         
-        assertInternalState()
+        calculator.input("+")
+        calculator.input(4)
         
         XCTAssertEqual(
             calculator.input(.instruction(.operator(.equal))),
             9
         )
-        
-        assertInternalState()
-        
-        XCTAssertEqual(
-            calculator.input(.instruction(.operator(.equal))),
-            12
-        )
-        
-        
-        XCTAssertEqual(calculator._value, 12)
+      
+        XCTAssertEqual(calculator._value, 9)
     }
     
     func test_addition() {
@@ -54,110 +42,79 @@ final class CalculatorTests: XCTestCase {
     }
     
     func test_instruction_equals_sub2() {
-        let calculator = Calculator._withState(
-            value: 15,
-            lastOperand: 5,
-            lastBinaryOperator: .subtaction
-        )
+        let calculator = Calculator()
        
-        func assertInternalState() {
-             XCTAssertEqual(calculator._lastOperand, 5)
-             XCTAssertEqual(calculator._lastBinaryOperator, .subtaction)
-             XCTAssertTrue(calculator._digits.isEmpty)
+        func initialState() {
+            calculator.input(2)
+            calculator.input("-")
+            calculator.input(3)
         }
         
-        assertInternalState()
+       initialState()
         XCTAssertEqual(
             calculator.input(.instruction(.operator(.equal))),
-            10
+            -1
         )
-        
-        assertInternalState()
-        XCTAssertEqual(
-            calculator.input(.instruction(.operator(.equal))),
-            5
-        )
-        
-        assertInternalState()
-        XCTAssertEqual(
-            calculator.input(.instruction(.operator(.equal))),
-            0
-        )
-        
-        assertInternalState()
+        calculator.input("-")
+        calculator.input(4)
         XCTAssertEqual(
             calculator.input(.instruction(.operator(.equal))),
             -5
         )
         
+        
         XCTAssertEqual(calculator._value, -5)
     }
     
     func test_instruction_equals_mul2() {
-        let calculator = Calculator._withState(
-            value: 2,
-            lastBinaryOperator: .multiplication
-        )
-       
-        func assertInternalState() {
-             XCTAssertEqual(calculator._lastOperand, 2)
-             XCTAssertEqual(calculator._lastBinaryOperator, .multiplication)
-             XCTAssertTrue(calculator._digits.isEmpty)
+        //Issues with using multiplication after equals
+        let calculator = Calculator()
+        
+        func initialState() {
+            calculator.input(2)
+            calculator.input("X")
+            calculator.input(3)
         }
         
-        assertInternalState()
+        initialState()
         XCTAssertEqual(
             calculator.input(.instruction(.operator(.equal))),
-            4
+            6
         )
         
-        assertInternalState()
+        calculator.input("X")
+        calculator.input(3)
+        
         XCTAssertEqual(
             calculator.input(.instruction(.operator(.equal))),
-            8
+            18
         )
         
-        assertInternalState()
-        XCTAssertEqual(
-            calculator.input(.instruction(.operator(.equal))),
-            16
-        )
-        
-        
-        XCTAssertEqual(calculator._value, 16)
+        XCTAssertEqual(calculator._value, 18)
     }
     
     func test_instruction_equals_div2() {
-        let calculator = Calculator._withState(
-            value: 512,
-            lastOperand: 2,
-            lastBinaryOperator: .division
-        )
+        //Currently throws and error, needs fixing.
+        let calculator = Calculator()
        
-        func assertInternalState() {
-             XCTAssertEqual(calculator._lastOperand, 2)
-             XCTAssertEqual(calculator._lastBinaryOperator, .division)
-             XCTAssertTrue(calculator._digits.isEmpty)
+        func initialState() {
+            calculator.input(512)
+            calculator.input(.instruction(.operator(.binaryOperator(.division))))
+            calculator.input(2)
         }
         
-        assertInternalState()
+        initialState()
         XCTAssertEqual(
             calculator.input(.instruction(.operator(.equal))),
             256
         )
         
-        assertInternalState()
-        XCTAssertEqual(
-            calculator.input(.instruction(.operator(.equal))),
-            128
-        )
-        
-        assertInternalState()
+        calculator.input(.instruction(.operator(.binaryOperator(.division))))
+        calculator.input(2)
         XCTAssertEqual(
             calculator.input(.instruction(.operator(.equal))),
             64
         )
-        
         
         XCTAssertEqual(calculator._value, 64)
     }
